requirement.txt
"""
torch>=1.12.0
torchvision>=0.13.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
Pillow>=8.3.0
python-json-logger>=2.0.0
"""